django/requirements.txt
Django==3.1.7
numpy==1.17
Pillow==8.3.2
torch==1.10.1
torchvision==0.11.2
transformers==4.18.0
mysqlclient==2.0.3
django-cors-headers==3.7.0
